niocortex_saas-main (3)/niocortex_saas-main/tas/requirements.txt
fastapi
uvicorn[standard]
pydantic-settings
sqlalchemy
asyncpg
aioredis
numpy
python-dotenv
sentence-transformers